+++ example/test/example_test.dart
@@ -1,6 +1,6 @@
 import 'package:flutter_test/flutter_test.dart';
 
-import 'package:example/main.dart';
+
 
 void main() {
   test('adds one to input values', () {